updated docs, moved get artifact helper to api utils
