+++ docs/user_guide/monitoring/metrics.rst
@@ -1,0 +1,6 @@
+Metrics
+===========
+
+.. automodule:: neptunecontrib.monitoring.metrics
+    :members:
+    :show-inheritance:
