--- docs/user_guide/monitoring/reporting.rst
@@ -1,6 +1,0 @@
-Reporting
-===========
-
-.. automodule:: neptunecontrib.monitoring.reporting
-    :members:
-    :show-inheritance:
